requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-jose
jinja2
cachetools
passlib
